# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AI Personal Health Companion API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
}

export type ProfileInputGender = typeof ProfileInputGender[keyof typeof ProfileInputGender];


export const ProfileInputGender = {
  male: 'male',
  female: 'female',
  other: 'other',
} as const;

export type ProfileInputConditionsItem = typeof ProfileInputConditionsItem[keyof typeof ProfileInputConditionsItem];


export const ProfileInputConditionsItem = {
  blood_pressure: 'blood_pressure',
  diabetes: 'diabetes',
  thyroid: 'thyroid',
  none: 'none',
} as const;

export interface ProfileInput {
  name?: string;
  /**
     * @minimum 1
     * @maximum 120
     */
  age: number;
  gender: ProfileInputGender;
  weightKg: number;
  conditions: ProfileInputConditionsItem[];
  medicationsText?: string;
  historyText?: string;
}

export interface ProfileResponse {
  id: string;
  userId: string;
  name: string;
  age: number;
  gender: string;
  weightKg: number;
  conditions: string[];
  medicationsText?: string;
  historyText?: string;
  createdAt: string;
  updatedAt: string;
}

export interface MedicationInput {
  name: string;
  dose: string;
  condition?: string;
  /**
     * @minimum 1
     * @maximum 8
     */
  timesPerDay: number;
  withFood?: boolean;
  active?: boolean;
}

export interface Medication {
  id: string;
  userId: string;
  name: string;
  dose: string;
  condition?: string;
  timesPerDay: number;
  withFood: boolean;
  active: boolean;
  createdAt: string;
}

export type MedicationLogInputStatus = typeof MedicationLogInputStatus[keyof typeof MedicationLogInputStatus];


export const MedicationLogInputStatus = {
  taken: 'taken',
  missed: 'missed',
  snoozed: 'snoozed',
} as const;

export interface MedicationLogInput {
  status: MedicationLogInputStatus;
  scheduledTime?: string;
}

export type MedicationLogStatus = typeof MedicationLogStatus[keyof typeof MedicationLogStatus];


export const MedicationLogStatus = {
  taken: 'taken',
  missed: 'missed',
  snoozed: 'snoozed',
} as const;

export interface MedicationLog {
  id: string;
  medicationId: string;
  userId: string;
  status: MedicationLogStatus;
  scheduledTime?: string;
  loggedAt: string;
}

export interface SymptomInput {
  /** @minLength 3 */
  inputText: string;
}

export type SymptomLogTier = typeof SymptomLogTier[keyof typeof SymptomLogTier];


export const SymptomLogTier = {
  home_care: 'home_care',
  monitor: 'monitor',
  see_doctor_24h: 'see_doctor_24h',
  emergency: 'emergency',
} as const;

export interface SymptomLog {
  id: string;
  userId: string;
  inputText: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  riskScore: number;
  tier: SymptomLogTier;
  aiReason: string;
  advice: string[];
  isEmergency: boolean;
  createdAt: string;
}

export type ReadingInputType = typeof ReadingInputType[keyof typeof ReadingInputType];


export const ReadingInputType = {
  bp: 'bp',
  sugar: 'sugar',
  weight: 'weight',
} as const;

export interface ReadingInput {
  type: ReadingInputType;
  /** For BP readings - systolic value */
  systolic?: number;
  /** For BP readings - diastolic value */
  diastolic?: number;
  /** For sugar (mg/dL) or weight (kg) readings */
  value?: number;
  note?: string;
}

export type ReadingType = typeof ReadingType[keyof typeof ReadingType];


export const ReadingType = {
  bp: 'bp',
  sugar: 'sugar',
  weight: 'weight',
} as const;

export interface Reading {
  id: string;
  userId: string;
  type: ReadingType;
  systolic?: number;
  diastolic?: number;
  value?: number;
  note?: string;
  createdAt: string;
}

export type FileUploadInputFileType = typeof FileUploadInputFileType[keyof typeof FileUploadInputFileType];


export const FileUploadInputFileType = {
  report: 'report',
  prescription: 'prescription',
} as const;

export interface FileUploadInput {
  fileType: FileUploadInputFileType;
  filename: string;
  title?: string;
  /** Base64-encoded file content */
  base64Data: string;
  mimeType: string;
}

export type HealthFileFileType = typeof HealthFileFileType[keyof typeof HealthFileFileType];


export const HealthFileFileType = {
  report: 'report',
  prescription: 'prescription',
} as const;

export interface HealthFile {
  id: string;
  userId: string;
  fileType: HealthFileFileType;
  filename: string;
  title?: string;
  ocrSummary?: string;
  storageKey?: string;
  uploadedAt: string;
}

export interface DailyRecommendation {
  id: string;
  userId: string;
  date: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  healthScore: number;
  greeting: string;
  fruitSuggestion: string;
  waterGoalLiters: number;
  sleepTargetHours: number;
  exerciseSuggestion: string;
  personalizedTip: string;
  dietPlan: {
    food: string;
    timing: string;
    benefit: string;
    imagePrompt: string;
  }[];
  warnings: string[];
  createdAt: string;
  cached: boolean;
  creditCost: {
    recommendation: number;
    aiTip: number;
    total: number;
  };
}

export interface CreditsResponse {
  used: number;
  limit: number;
  remaining: number;
  resetDate: string;
}

export type GetMedications200 = {
  medications: Medication[];
};

export type GetMedicationLogs200 = {
  logs: MedicationLog[];
};

export type GetSymptomLogsParams = {
limit?: number;
};

export type GetSymptomLogs200 = {
  logs: SymptomLog[];
};

export type GetReadingsParams = {
type?: GetReadingsType;
limit?: number;
};

export type GetReadingsType = typeof GetReadingsType[keyof typeof GetReadingsType];


export const GetReadingsType = {
  bp: 'bp',
  sugar: 'sugar',
  weight: 'weight',
} as const;

export type GetReadings200 = {
  readings: Reading[];
};

export type GetFilesParams = {
type?: GetFilesType;
};

export type GetFilesType = typeof GetFilesType[keyof typeof GetFilesType];


export const GetFilesType = {
  report: 'report',
  prescription: 'prescription',
} as const;

export type GetFiles200 = {
  files: HealthFile[];
};