# lib/api-zod/src/generated/types/medicationInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AI Personal Health Companion API
 * OpenAPI spec version: 0.1.0
 */

export interface MedicationInput {
  name: string;
  dose: string;
  condition?: string;
  /**
     * @minimum 1
     * @maximum 8
     */
  timesPerDay: number;
  withFood?: boolean;
  active?: boolean;
}