# lib/api-zod/src/generated/types/medicationLogInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AI Personal Health Companion API
 * OpenAPI spec version: 0.1.0
 */
import type { MedicationLogInputStatus } from './medicationLogInputStatus';

export interface MedicationLogInput {
  status: MedicationLogInputStatus;
  scheduledTime?: Date;
}