# lib/api-zod/src/generated/types/profileResponse.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AI Personal Health Companion API
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileResponse {
  id: string;
  userId: string;
  age: number;
  gender: string;
  weightKg: number;
  conditions: string[];
  medicationsText?: string;
  historyText?: string;
  createdAt: Date;
  updatedAt: Date;
}