# lib/api-zod/src/generated/types/reading.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AI Personal Health Companion API
 * OpenAPI spec version: 0.1.0
 */
import type { ReadingType } from './readingType';

export interface Reading {
  id: string;
  userId: string;
  type: ReadingType;
  systolic?: number;
  diastolic?: number;
  value?: number;
  note?: string;
  createdAt: Date;
}